restructured files, added router
