+++ src/Components/Navbar.js
@@ -1,4 +1,4 @@
-import logo from './logo.svg';
+import logo from '../Assets/Images/logo.svg';
 import './Navbar.css'
 
 function Navbar() {
