+++ src/Pages/Events.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Events() {
+    return (
+        <div>
+            Hello World
+        </div>
+    )
+}
+
+export default Events;
